# Backend/openapi.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/ulfix9/api-gersa/1.0.0
  - description: server
    url: http://localhost:4000/api
info: 
  title: API GERSA
  version: 1.0.0
...
paths: 
...
  /employees/{employeeId}:
    patch:
      tags: 
      - employees
      summary: End Point para modificar los atributos del empleado
      parameters:
      - name: employeeId
        in: path
        description: Identificador del empleado
        required: true
        schema:
          type: integer
      requestBody: 
        description: "Para realizar un patch no es necesario enviar todo el schema, solo el atributo a modificar."
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaEmployee'
      responses:
        200:
          description: Actualización de datos correctamente.
          $ref: '#/components/schemas/Exito'
        400:
          $ref: '#/components/schemas/BadRequest'
        401:
          $ref: '#/components/schemas/UnauthorizedError'
        404:
          $ref: '#/components/schemas/NotFound'
        500:
          $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
    
    put:
      tags: 
      - employees
      summary: End Point para modificar los atributos del empleado
      parameters:
      - name: employeeId
        in: path
        description: Identificador del empleado
        required: true
        schema:
          type: integer
      requestBody: 
        description: "Para realizar un put es necesario enviar todo el schema del empleado."
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaEmployee'
      responses:
        200:
          description: Actualización de datos correctamente.
          $ref: '#/components/schemas/Exito'
        400:
          $ref: '#/components/schemas/BadRequest'
        401:
          $ref: '#/components/schemas/UnauthorizedError'
        404:
          $ref: '#/components/schemas/NotFound'
        500:
          $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
    get:
      tags: 
      - employees
      summary: End Point para obtener los atributos del empleado
      parameters:
      - name: employeeId
        in: path
        description: Identificador del empleado
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Peticion exitosa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaEmployee'
        400:
          $ref: '#/components/schemas/BadRequest'
        401:
          $ref: '#/components/schemas/UnauthorizedError'
        404:
          $ref: '#/components/schemas/NotFound'
        500:
          $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
...
  /employees/{employeeId}/hoursAccepted:
    get:
      tags:
        - "employees"
      summary: Obtiene la información de las horas extra aceptadas por el jefe de cuadrilla
      parameters:
      - name: employeeId
        in: path
        description: Identificador del empleado
        required: true
        schema:
          type: integer
      responses:
        200:
          description: ok
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/hoursAcepted'
        202:
          description: bad
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/hourNotFound'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
    post:
        tags:
        - employees
        summary: End Point para el registro de horas extra aceptadas
        description: El End Point recibe por body el nombre del tipo de evento
        requestBody: 
          description: "Para realizar un post es necesario enviar todo el schema."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodyHourAcept'
        responses:
          201:
            description: (OK) La informacion de los contratos se obtuvo de forma exitosa
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/hoursAcepted'
          400:
            description: Bad Request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/BadRequest'
          401:
            description: Unauthorized
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/UnauthorizedError'
          403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
          500:
            description: Server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ServerError'
        security:
          - bearerAuth: []
  /employees/{employeeId}/timeRequest:
    get:
      tags:
      - employees
      summary: Obtiene la informacion de la base de datos de las solicitudes de tiempo extra de un empleado
      parameters:
      - name: employeeId
        in: path
        description: Identificador de la solicitud a obtener
        required: true
        schema:
          type: integer
      responses:
        200:
          description: (OK) La informacion de las solicitudes se obtuvo de forma exitosa 
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/ExitoSolicitudTiempoGet'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
    post:
      tags:
      - employees
      summary: Registra la solicitud de tiempo extra de un empleado
      description: "El EndPoint registra la solicitud de tiempo extra "
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: number
      responses:
        201:
          description: ok
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/TimeRequestPOST'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeRequestsRegister"
      security:
        - bearerAuth: []
  /employees/timeRequest/:
   get:
      tags:
      - employees
      summary: Obtiene la informacion de la base de datos de las solicitudes de tiempo extra de un empleado, de varios empleados o todos los empleados
      responses:
        200:
          description: (OK) La informacion de las solicitudes de tiempo extra se obtuvo de forma exitosa
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/ExitoSolicitudTiempoGet'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []     
  /employees/timeRequest/{timeRequestId}:
    patch:
      tags:
      - employees
      summary: Modifica datos en especifico del contrato 
      description: "El EndPoint recibe via url el id de la solicitud de tiempo extra que se desea modificar, en caso de ser exitoso se regresan los datos modificados de la solicitud. Para realizar la petición se necesita mandar token por headers"
      parameters:
      - name: timeRequestId
        in: path
        description: Identificador del empleado
        required: true
        schema:
          type: integer
      requestBody: 
        description: "Para realizar un patch no es necesario enviar todo el schema, solo el atributo a modificar."
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodySolicitudesTiempoPatch'
        required: false
      responses:
        201:
          description: (OK), modificación exitosa 
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/ExitoSolicitudTiempoPatch'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
  /employees/{employeeId}/reports:
    get:
      tags:
      - employees
      summary: Obtiene la informacion de la base de datos de los reportes de un empleado
      parameters:
      - name: employeeId
        in: path
        description: Identificador de las solicitudes a obtener
        required: true
        schema:
          type: integer
      responses:
        200:
          description: (OK) La informacion de las solicitudes se obtuvo de forma exitosa
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/ReportsRegister"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []  
    post:
      tags:
      - employees
      summary: Registra el reporte del empleado seleccionado
      description: "El EndPoint registra el reporte del empleado"
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: number
      responses:
        201:
          description: ok
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/ReportsPOST'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportsRegister"
      security:
        - bearerAuth: []
  /employees/reports/:
   get:
      tags:
      - employees
      summary: Obtiene la informacion de la base de datos de los reportes de todos los empleados
      responses:
        200:
          description: (OK) La informacion de los reportes se obtuvo de forma exitosa
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/ExitoReportesGet'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        403:
            description: Forbidden
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        500:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
      security:
        - bearerAuth: []
...
  /employees/{userId}/accessCode:
    get:
      tags:
      - employees
      summary: Recupera el código de acceso
      description: "El EndPoint recibe via url el id del usuario"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        200:
          description: (OK), modificación exitosa 
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/Exito'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - bearerAuth: []
...
components: 
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
  schemas:
...
    SchemaEmployee:
      type: object
      properties:
        message:
         type: string
         description: Completado con exito
        ok:
         type: boolean
         description: (True) Salio bien, (False) Algo salio mal
        data: 
          type: object
          properties:
            tipoIdentificacion:
              type: string
            documentoIdentidad:
              type: string
            fechaNacimiento: 
              type: string
              format: date
            nacionalidad: 
              type: string
            genero: 
              type: string
            lugarDeTrabajo: 
              type: string
            supervisor:
              type: string
            numeroCuentaBancaria:
              type: string
            swiftBic:
              type: string
            frecuenciaPago: 
              type: string
            direccion1: 
              type: string
            direccion2:
              type: string
            ciudad: 
              type: string
            codigoPostal: 
              type: string
            estadoProvincia: 
              type: string
            pais: 
              type: string
            emergenciaNombre: 
              type: string
            empergenciaTelefono: 
              type: string
            rfc:
              type: string
            numeroImms: 
              type: string
            curp: 
              type: string
            fechaAltaImss:
              type: string
              format: date
            scheduleId:
              type: integer
...